BE: Move stuff around to the correct places
