+++ backend/api/src/main/kotlin/io/tolgee/api/v2/controllers/NotificationController.kt
@@ -1,11 +1,11 @@
-package io.tolgee.ee.api.v2.controllers
+package io.tolgee.api.v2.controllers
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import io.tolgee.hateoas.notification.NotificationResponse
 import io.tolgee.security.authentication.AllowApiAccess
 import io.tolgee.security.authentication.AuthenticationFacade
-import io.tolgee.service.NotificationModel
-import io.tolgee.service.NotificationService
+import io.tolgee.service.notification.NotificationService
 import org.springframework.web.bind.annotation.CrossOrigin
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -19,21 +19,17 @@ import org.springframework.web.bind.annotation.RestController
   ],
 )
 @Tag(name = "Notifications", description = "Manipulates notifications")
-class NotificationsController(
+class NotificationController(
   private val notificationService: NotificationService,
   private val authenticationFacade: AuthenticationFacade,
 ) {
 
   @GetMapping
-  @Operation(summary = "Get notifications")
+  @Operation(summary = "Gets notifications of the currently logged in user, newest is first.")
   @AllowApiAccess
-  fun getNotifications(): NotificationsResponse {
-    return NotificationsResponse(
+  fun getNotifications(): NotificationResponse {
+    return NotificationResponse(
       notificationService.getNotifications(authenticationFacade.authenticatedUser.id)
     )
   }
 }
-
-data class NotificationsResponse(
-  val notifications: List<NotificationModel>,
-)
