+++ backend/api/src/main/kotlin/io/tolgee/hateoas/notification/NotificationResponse.kt
@@ -1,0 +1,7 @@
+package io.tolgee.hateoas.notification
+
+import io.tolgee.service.notification.NotificationModel
+
+data class NotificationResponse(
+  val notifications: List<NotificationModel>,
+)
