+++ backend/data/src/main/kotlin/io/tolgee/service/notification/NotificationModel.kt
@@ -1,0 +1,8 @@
+package io.tolgee.service.notification
+
+data class NotificationModel(
+  val id: Long,
+  val linkedProjectId: Long?,
+  val linkedTaskNumber: Long?,
+  val linkedTaskName: String?,
+)
